backend/requirements/new.txt
aerich==0.7.2
aiosqlite==0.17.0
anyio==4.6.0
asgiref==3.8.1
asyncpg==0.25.0
bcrypt==4.2.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
cryptography==43.0.1
dictdiffer==0.9.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.68.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
iniconfig==2.0.0
iso8601==0.1.16
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
pyasn1==0.6.1
pycparser==2.22
pydantic==1.10.18
pypika-tortoise==0.1.6
pytest==7.3.1
pytest-asyncio==0.21.0
python-dotenv==0.19.0
python-jose==3.3.0
python-multipart==0.0.12
pytz==2024.2
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.14.2
tomli==2.0.1
tomlkit==0.13.2
tortoise-orm==0.17.8
typing_extensions==4.12.2
uvicorn==0.15.0
